requirements.txt
backports.weakref==1.0rc1
bleach==1.5.0
html5lib==0.9999999
Markdown==2.6.8
numpy==1.13.1
olefile==0.44
Pillow==4.2.1
pkg-resources==0.0.0
protobuf==3.3.0
scikit-learn==0.18.2
scipy==0.19.1
six==1.10.0
sklearn==0.0
tensorflow-gpu==1.2.1
Werkzeug==0.12.2
